Post Call from the Android Application
I cannot remove the token field from the post call and moreover im
getting issues with the json accepting int values.

The token is generated when passing all the values as strings.

as of now , some of the values are still hard-coded.

Fixed some minor issues with the app. Please test and raise a defect for
all the issues.

added the generated token dialogue code as popup.